+++ programs-ecs/bundles/ecs-bundle/Cargo.toml
@@ -14,7 +14,7 @@ no-idl = []
 no-log-ix-name = []
 cpi = ["no-entrypoint"]
 default = []
-idl-build = ["bolt-lang/idl-build"]
+idl-build = ["bolt-lang/idl-build", "anchor-lang/idl-build", "anchor-spl/idl-build"]
 anchor-debug = ["bolt-lang/anchor-debug"]
 custom-heap = []
 custom-panic = []
@@ -22,4 +22,7 @@ custom-panic = []
 
 [dependencies]
 bolt-lang.workspace = true
-serde = { version = "1.0", features = ["derive"] }
+anchor-lang.workspace = true
+anchor-spl.workspace = true
+serde.workspace = true
+token-minter.workspace = true
