+++ programs-ecs/bundles/ecs-bundle/src/systems/exchange/mod.rs
@@ -1,0 +1,11 @@
+use bolt_lang::*;
+
+#[error_code]
+pub enum ExchangeError {
+    #[msg("No Exchange")]
+    NoExchange,
+    #[msg("Not Enough Tokens")]
+    NotEnoughTokens,
+    #[msg("Token Burn Failed")]
+    TokenBurnFailed,
+}
