--- programs-ecs/systems/bump_location_allocator/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "bump_location_allocator"
-version = "0.1.8"
-description = "Created with Bolt"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "lib"]
-name = "bump_location_allocator"
-
-[features]
-no-entrypoint = []
-no-idl = []
-no-log-ix-name = []
-cpi = ["no-entrypoint"]
-default = []
-idl-build = ["anchor-lang/idl-build"]
-
-[dependencies]
-bolt-lang.workspace = true
-anchor-lang.workspace = true
-serde.workspace = true
-ecs-bundle.workspace = true
