--- programs-ecs/systems/claim-loot/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "claim-loot"
-version = "0.1.8"
-description = "Created with Bolt"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "lib"]
-name = "claim_loot"
-
-[features]
-no-entrypoint = []
-no-idl = []
-no-log-ix-name = []
-cpi = ["no-entrypoint"]
-default = []
-idl-build = ["anchor-lang/idl-build", "anchor-spl/idl-build"]
-
-[dependencies]
-bolt-lang.workspace = true
-anchor-lang.workspace = true
-anchor-spl.workspace = true
-serde.workspace = true
-ecs-bundle.workspace = true
-token-minter = { version = "0.1.0", path = "../../../programs/token-minter", features = ["cpi"] }
