--- programs-ecs/systems/claim-time/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "claim-time"
-version = "0.1.10"
-description = "Created with Bolt"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "lib"]
-name = "claim_time"
-
-[features]
-no-entrypoint = []
-no-idl = []
-no-log-ix-name = []
-cpi = ["no-entrypoint"]
-default = []
-idl-build = ["anchor-lang/idl-build"]
-
-[dependencies]
-bolt-lang.workspace = true
-anchor-lang.workspace = true
-serde.workspace = true
-ecs-bundle.workspace = true
