+++ programs-ecs/systems/exchange/Cargo.toml
@@ -14,12 +14,12 @@ no-idl = []
 no-log-ix-name = []
 cpi = ["no-entrypoint"]
 default = []
-idl-build = ["anchor-lang/idl-build", "anchor-spl/idl-build"]
+idl-build = ["anchor-lang/idl-build", "anchor-spl/idl-build", "ecs-bundle/idl-build"]
 
 [dependencies]
 bolt-lang.workspace = true
 anchor-lang.workspace = true
 serde.workspace = true
 anchor-spl.workspace = true
-token-minter = { version = "0.1.0", path = "../../../programs/token-minter", features = ["cpi"] }
+token-minter.workspace = true
 ecs-bundle.workspace = true
