+++ programs-ecs/systems/repair/Cargo.toml
@@ -14,7 +14,7 @@ no-idl = []
 no-log-ix-name = []
 cpi = ["no-entrypoint"]
 default = []
-idl-build = ["anchor-lang/idl-build"]
+idl-build = ["anchor-lang/idl-build", "ecs-bundle/idl-build"]
 
 [dependencies]
 bolt-lang.workspace = true
